requirements.txt
favicon>=0.5.1
feedparser>=5.2.1
Pillow>=5.4.0
yattag>=1.10.1
arrow>=0.12.1
inflect>=2.1.0
python-slugify>=2.0.1
bleach>=3.0.2
